Dataset: Add dataset.DatasetLibrary.
