+++ src/pymovements/dataset/dataset_definition.py
@@ -1,0 +1,86 @@
+# Copyright (c) 2023 The pymovements Project Authors
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+"""DatasetDefinition module."""
+from __future__ import annotations
+
+from dataclasses import dataclass
+from dataclasses import field
+from typing import Any
+
+from pymovements.gaze.experiment import Experiment
+
+
+@dataclass
+class DatasetDefinition:
+    """Definition to initialize a :py:class:`~pymovements.Dataset`.
+
+    Attributes
+    ----------
+    name : str
+        The name of the dataset.
+
+    mirrors : tuple[str, ...]
+        A tuple of mirrors of the dataset. Each entry must be of type `str` and end with a '/'.
+
+    resources : tuple[dict[str, str], ...]
+        A tuple of dataset resources. Each list entry must be a dictionary with the following keys:
+        - `resource`: The url suffix of the resource. This will be concatenated with the mirror.
+        - `filename`: The filename under which the file is saved as.
+        - `md5`: The MD5 checksum of the respective file.
+
+    experiment : Experiment
+        The experiment definition.
+
+    filename_regex : str
+        Regular expression which will be matched before trying to load the file. Namedgroups will
+        appear in the `fileinfo` dataframe.
+
+    filename_regex_dtypes : dict[str, type], optional
+        If named groups are present in the `filename_regex`, this makes it possible to cast specific
+        named groups to a particular datatype.
+
+    column_map : dict[str, str]
+        The keys are the columns to read, the values are the names to which they should be renamed.
+
+    custom_read_kwargs : dict[str, Any], optional
+        If specified, these keyword arguments will be passed to the file reading function.
+    """
+    name: str = field(init=False)
+
+    mirrors: tuple[str, ...] = field(default_factory=tuple)
+
+    resources: tuple[dict[str, str], ...] = field(default_factory=tuple)
+
+    experiment: Experiment | None = None
+
+    filename_regex: str = '.*'
+
+    filename_regex_dtypes: dict[str, type] = field(default_factory=dict)
+
+    custom_read_kwargs: dict[str, Any] = field(default_factory=dict)
+
+    column_map: dict[str, str] = field(default_factory=dict)
+
+    def __post_init__(self):
+        self.name = self.__class__.__name__
+
+        if len(self.column_map) > 0:
+            self.custom_read_kwargs['columns'] = list(self.column_map.keys())
+            self.custom_read_kwargs['new_columns'] = list(self.column_map.values())
