+++ src/pymovements/datasets/gazebase.py
@@ -17,29 +17,36 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-"""This module provides an interface to the JuDo1000 dataset."""
+"""This module provides an interface to the GazeBase dataset."""
 from __future__ import annotations
 
 from dataclasses import dataclass
 from dataclasses import field
+from typing import Any
 
-from pymovements.datasets.public_dataset import PublicDatasetDefinition
-from pymovements.datasets.public_dataset import register_public_dataset
+from pymovements.dataset.dataset_definition import DatasetDefinition
+from pymovements.dataset.dataset_library import register_dataset
 from pymovements.gaze.experiment import Experiment
 
 
 @dataclass
-@register_public_dataset
-class JuDo1000(PublicDatasetDefinition):
-    """JuDo1000 dataset :cite:p:`JuDo1000`.
+@register_dataset
+class GazeBase(DatasetDefinition):
+    """GazeBase dataset :cite:p:`GazeBase`.
 
-    This dataset includes binocular eye tracking data from 150 participants in four sessions with an
-    interval of at least one week between two sessions. Eye movements are recorded at a sampling
-    frequency of 1000 Hz using an EyeLink Portable Duo video-based eye tracker and are provided as
-    pixel coordinates. Participants are instructed to watch a random jumping dot on a computer
-    screen.
+    This dataset includes monocular (left eye) eye tracking data from 322 participants captured over
+    a period of 37 months. Participants attended up to 9 rounds during this time frame, with each
+    round consisting of two contiguous sessions.
 
-    Check the respective `repository <https://osf.io/5zpvk/>`_ for details.
+    Eye movements are recorded at a sampling frequency of 1000 Hz using an EyeLink 1000 video-based
+    eye tracker and are provided as positional data in degrees of visual angle.
+
+    In each of the two sessions per round, participants are instructed to complete a series of
+    tasks, including a fixation task (FIX), a horizontal saccade task (HSS), a random saccade task
+    (RAN), a reading task (TEX), two free viewing video tasks (VD1 and VD2) and a gaze-driven gaming
+    task (BLG).
+
+    Check the respective paper for details :cite:p:`GazeBase`.
 
     Attributes
     ----------
@@ -75,11 +82,11 @@ class JuDo1000(PublicDatasetDefinition):
     Examples
     --------
     Initialize your :py:class:`~pymovements.PublicDataset` object with the
-    :py:class:`~pymovements.JuDo1000` definition:
+    :py:class:`~pymovements.GazeBase` definition:
 
     >>> import pymovements as pm
     >>>
-    >>> dataset = pm.PublicDataset("JuDo1000", root='data/')
+    >>> dataset = pm.PublicDataset("GazeBase", root='data/')
 
     Download the dataset resources resources:
 
@@ -92,34 +99,39 @@ class JuDo1000(PublicDatasetDefinition):
     # pylint: disable=similarities
     # The PublicDatasetDefinition child classes potentially share code chunks for definitions.
 
-    name: str = 'JuDo1000'
+    name: str = 'GazeBase'
 
-    mirrors: tuple[str, ...] = (
-        'https://osf.io/download/',
+    mirrors: tuple[str] = (
+        'https://figshare.com/ndownloader/files/',
     )
 
-    resources: tuple[dict[str, str], ...] = (
+    resources: tuple[dict[str, str]] = (
         {
-            'resource': '4wy7s/',
-            'filename': 'JuDo1000.zip',
-            'md5': 'b8b9e5bb65b78d6f2bd260451cdd89f8',
+            'resource': '27039812',
+            'filename': 'GazeBase_v2_0.zip',
+            'md5': 'cb7eb895fb48f8661decf038ab998c9a',
         },
     )
 
     experiment: Experiment = Experiment(
-        screen_width_px=1280,
-        screen_height_px=1024,
-        screen_width_cm=38,
-        screen_height_cm=30.2,
-        distance_cm=68,
+        screen_width_px=1680,
+        screen_height_px=1050,
+        screen_width_cm=47.4,
+        screen_height_cm=29.7,
+        distance_cm=55,
         origin='lower left',
         sampling_rate=1000,
     )
 
-    filename_regex: str = r'(?P<subject_id>\d+)_(?P<session_id>\d+).csv'
+    filename_regex: str = (
+        r'S_(?P<round_id>\d)(?P<subject_id>\d+)'
+        r'_S(?P<session_id>\d+)'
+        r'_(?P<task_name>.+).csv'
+    )
 
     filename_regex_dtypes: dict[str, type] = field(
         default_factory=lambda: {
+            'round_id': int,
             'subject_id': int,
             'session_id': int,
         },
@@ -127,18 +139,13 @@ class JuDo1000(PublicDatasetDefinition):
 
     column_map: dict[str, str] = field(
         default_factory=lambda: {
-            'trialId': 'trial_id',
-            'pointId': 'point_id',
-            'time': 'time',
-            'x_left': 'x_left_pix',
-            'y_left': 'y_left_pix',
-            'x_right': 'x_right_pix',
-            'y_right': 'y_right_pix',
+            'n': 'time',
+            'x': 'x_left_pos',
+            'y': 'y_left_pos',
+            'val': 'validity',
+            'xT': 'x_target_pos',
+            'yT': 'y_target_pos',
         },
     )
 
-    custom_read_kwargs: dict[str, str] = field(
-        default_factory=lambda: {
-            'separator': '\t',
-        },
-    )
+    custom_read_kwargs: dict[str, Any] = field(default_factory=dict)
